build-and-provide-package: rework package removal for arch-all packages
We can't remove a package from the repository in the step where
we're checking for missing binary packages (to get rid of packages
which are no longer provided but shouldn't stay in the archive)
if the package is an 'Architecture: all' one, as e.g. the i386
build might remove the package from the amd64 build.

So let's check for "-B" in DEBBUILDOPTS and then skip the removal
process in such a situation...